Update reactive-style; renamed to style-it to match api
