+++ README.md
@@ -54,11 +54,11 @@ For example, if a package supports the css file extraction you can run the autop
 | [react-theme](https://github.com/azazdeaz/react-theme) | 0.1.4 | | | | x | |
 | [react-vstyle](https://github.com/fdecampredon/react-vstyle) | 0.1.0 | x | x | x | x | x |
 | [reactcss](https://github.com/casesandberg/reactcss) | 0.3.2 | x | | | x | |
-| [reactive-style](https://github.com/buildbreakdo/reactive-style) | 1.0.0 | | x | x | | |
 | [scope-styles](https://github.com/rtsao/scope-styles) | 0.6.0 |  | x | x | x | x |
 | [smart-css](https://github.com/hackhat/smart-css) | 1.1.1 | | x | x | x | |
 | [stile](https://github.com/bloodyowl/stile) + [react-media-queries](https://github.com/bloodyowl/react-media-queries) | 2.0.0 | x | | x | x | | |
 | [stilr](https://github.com/kodyl/stilr) | 1.1.0 | | x | x | x | x |
+| [style-it](https://github.com/buildbreakdo/style-it) | 1.2.1 | | x | x | | |
 | [styling](https://github.com/andreypopp/styling) | 0.2.0 | | x | | x | x |
 
 ## Testimonials
