--- reactive-style/button.js
@@ -1,45 +1,0 @@
-import React, {Component} from 'react';
-import ReactDOM from 'react-dom';
-
-import Style from 'reactive-style';
-
-class Button extends Component {
-  render() {
-    return (
-      <Style>
-        {`
-          .container {
-            text-align: center;
-          }
-          .button {
-            background-color: #ff0000;
-            width: 320px;
-            padding: 20px;
-            border-radius: 5px;
-            border: none;
-            outline: none;
-          }
-          .button:hover {
-            color: #fff;
-          }
-          .button:active {
-            position: relative;
-            top: 2px;
-          }
-          @media (max-width: 480px) {
-            .button {
-              width: 160px;
-            }
-          }
-        `}
-
-        <div className="container">
-          <button className="button">Click me!</button>
-        </div>
-      </Style>
-    );
-  }
-
-}
-
-ReactDOM.render(<Button />, document.getElementById('content'));
