+++ style-it/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Reactive Style - CSS in JS</title>
+    <title>Style It - CSS in JS</title>
   </head>
   <meta charset="utf-8" />
   <body>
