+++ style-it/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "reactive-style-css-in-js",
+  "name": "style-it-css-in-js",
   "version": "1.0.0",
   "description": "Reactive Style - CSS in JS",
   "scripts": {
@@ -10,7 +10,7 @@
   "dependencies": {
     "react": "^15.3.0",
     "react-dom": "^15.3.0",
-    "reactive-style": "1.0.0"
+    "style-it": "*"
   },
   "babel": {
     "presets": ["es2015", "stage-1", "react"]
