+++ style-it/webpack.config.js
